requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
attrs==24.3.0
base58==2.1.1
bitarray==2.9.3
certifi==2024.12.14
charset-normalizer==3.4.0
ckzg==2.0.1
cytoolz==0.12.3
decorator==5.2.1
eth-account==0.13.5
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.1.0
eth-utils==5.2.0
eth_abi==5.2.0
ffmpeg-python==0.2.0
frozenlist==1.5.0
future==1.0.0
hexbytes==1.3.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
ipfshttpclient==0.8.0a2
jsonschema==3.2.0
lru-dict==1.3.0
moviepy==2.1.2
multiaddr==0.0.9
multidict==6.1.0
netaddr==1.3.0
numpy==2.2.3
parsimonious==0.10.0
pillow==10.4.0
proglog==0.1.10
propcache==0.2.1
protobuf==3.20.3
pycryptodome==3.21.0
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
pyrsistent==0.20.0
python-dotenv==1.0.1
pyunormalize==16.0.0
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
six==1.17.0
toolz==1.0.0
tqdm==4.67.1
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.2.3
varint==1.0.2
web3==7.8.0
websockets==13.1
yarl==1.18.3
